refactor: init constants, utils, make small structure refactoring
