+++ src/controller/GameController.js
@@ -1,0 +1,8 @@
+import Phaser from 'phaser';
+import {gameConfig} from '@controller/game/GameConfig';
+
+export default class GameController {
+  start() {
+    this.game = new Phaser.Game(gameConfig); // Start Phaser
+  }
+}
