+++ src/index.html
@@ -7,10 +7,6 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>RS Clone</title>
     <link rel="icon" href="favicon.ico" type="image/x-icon">
-    
 </head>
-<body>
-    <div id = "game"></div>
-</body>
-<script src="./index.js"></script>
+<body></body>
 </html>
